Se rodou tá valendo... Criado a rotina de registro mensal e o dashboard

+++ src/main/java/br/com/smarttech/frigonix/controllers/dtos/ResumoResponseRecordDTO.java
@@ -7,19 +7,19 @@ public record ResumoResponseRecordDTO(
     TaxaFalha falhas
 ) {
     public record Equipamentos(
-            Integer quantidadeAtual,
+            Long quantidadeAtual,
             Double vsMesAnterior
     ){}
     public record ManutencoesPendentes(
-            Integer quantidadePendentes,
-            Integer quantidadeUrgentes
+            Long quantidadePendentes,
+            Long quantidadeUrgentes
     ){}
     public record ManutencoesConcluidas(
-            Integer quantidadeAtual,
+            Long quantidadeAtual,
             Double vsMesAnterior
     ){}
     public record TaxaFalha(
-            Double quantidadeFalha,
+            Long quantidadeFalha,
             Double vsMesAnterior
     ){}
 }
